markdownConverter대신 depthCalculator을 작성하라
