+++ src/depthCalculator.js
@@ -1,0 +1,10 @@
+export default function depthCalcutaor(deletedTasks) {
+  return deletedTasks.reduceRight((acc, cur) => {
+    const { selfId, parentId } = cur;
+
+    return {
+      ...acc,
+      [selfId]: acc[parentId] + 1 || 1,
+    };
+  }, {});
+}
